ArbitrarilyOrientedRegularGrid - Issue 2969 (#2974)
* inc

* inc

* inc

* add new files

* inc

* clang-format

* clang-format

* add file

* clang-format

* clang-format

* inc

* inc

* cleanup

* strange git behavior...?

* document planeDescription

* delete unused versions of GetUserCoordinates

* remove perimeter sampling offset from ArbitrarilyOrientedRegularGrid, and move the SliceRenderer's winding-order and planar extents from the Grid class to the SliceRenderer

* delete functions that use CoordType

* add offset to _textureSideSize

* clang-format

* use raw pointer allocated in initializer list

* cleanup

* inc

* update grid constructor to current branch

+++ include/vapor/SliceRenderer.h
@@ -74,8 +74,8 @@ private:
     void _resetState();
     void _initializeState();
 
-    bool _initialized;
-    int  _textureSideSize;
+    bool   _initialized;
+    size_t _textureSideSize;
 
     GLuint _colorMapTextureID;
     GLuint _dataValueTextureID;
